Fixed Alert, now working

+++ Server/ServerApp/src/main/java/application/repositories/AlertRepository.java
@@ -7,6 +7,6 @@ import application.model.Alert;
 //This will be AUTO IMPLEMENTED by Spring into a Bean called alertRepository
 //CRUD refers Create, Read, Update, Delete
 
-public interface AlertRepository extends CrudRepository<Alert, Long> {
+public interface AlertRepository extends CrudRepository<Alert, Integer> {
 
 }
